feat/#14: 카테고리 추가 모달창 컴포넌트 분리
